third_party/blink/web_tests/accessibility/accessibility-node-appendChild-to-self-expected.txt

CONSOLE ERROR: Uncaught InvalidAccessError: Failed to execute 'appendChild' on 'AccessibleNode': An AccessibleNode cannot be a child of itself
